ref(new-widget-builder-experience): Move widget builder step components to separate files
