+++ static/app/views/dashboardsV2/widgetBuilder/buildSteps/yAxisStep/yAxisSelector/addButton.tsx
