+++ static/app/views/dashboardsV2/widgetBuilder/buildSteps/yAxisStep/yAxisSelector/deleteButton.tsx
