+++ static/app/views/dashboardsV2/widgetBuilder/widgetBuilder.tsx
@@ -8,23 +8,13 @@ import set from 'lodash/set';
 
 import {validateWidget} from 'sentry/actionCreators/dashboards';
 import {addErrorMessage, addSuccessMessage} from 'sentry/actionCreators/indicator';
-import Button from 'sentry/components/button';
 import {generateOrderOptions} from 'sentry/components/dashboards/widgetQueriesForm';
-import SearchBar from 'sentry/components/events/searchBar';
-import Input from 'sentry/components/forms/controls/input';
-import RadioGroup from 'sentry/components/forms/controls/radioGroup';
-import Field from 'sentry/components/forms/field';
-import SelectControl from 'sentry/components/forms/selectControl';
 import * as Layout from 'sentry/components/layouts/thirds';
-import ExternalLink from 'sentry/components/links/externalLink';
 import List from 'sentry/components/list';
 import LoadingError from 'sentry/components/loadingError';
 import PageFiltersContainer from 'sentry/components/organizations/pageFilters/container';
-import {PanelAlert} from 'sentry/components/panels';
 import SentryDocumentTitle from 'sentry/components/sentryDocumentTitle';
-import {MAX_QUERY_LENGTH} from 'sentry/constants';
-import {IconAdd, IconDelete} from 'sentry/icons';
-import {t, tct} from 'sentry/locale';
+import {t} from 'sentry/locale';
 import {PageContent} from 'sentry/styles/organization';
 import space from 'sentry/styles/space';
 import {
@@ -44,9 +34,7 @@ import {
   QueryFieldValue,
 } from 'sentry/utils/discover/fields';
 import handleXhrErrorResponse from 'sentry/utils/handleXhrErrorResponse';
-import Measurements, {
-  MeasurementCollection,
-} from 'sentry/utils/measurements/measurements';
+import {MeasurementCollection} from 'sentry/utils/measurements/measurements';
 import {SessionMetric} from 'sentry/utils/metrics/fields';
 import {SPAN_OP_BREAKDOWN_FIELDS} from 'sentry/utils/performance/spanOperationBreakdowns/constants';
 import useApi from 'sentry/utils/useApi';
@@ -66,41 +54,30 @@ import {
   WidgetQuery,
   WidgetType,
 } from 'sentry/views/dashboardsV2/types';
-import {
-  generateIssueWidgetFieldOptions,
-  generateIssueWidgetOrderOptions,
-} from 'sentry/views/dashboardsV2/widgetBuilder/issueWidget/utils';
 import {generateFieldOptions} from 'sentry/views/eventsV2/utils';
 import {IssueSortOptions} from 'sentry/views/issueList/utils';
 
 import {DEFAULT_STATS_PERIOD} from '../data';
-import WidgetCard from '../widgetCard';
 
-import BuildStep from './buildStep';
-import {ColumnFields} from './columnFields';
-import {DashboardSelector} from './dashboardSelector';
-import {DisplayTypeSelector} from './displayTypeSelector';
+import {ColumnsStep} from './buildSteps/columnsStep';
+import {DashboardStep} from './buildSteps/dashboardStep';
+import {DataSetStep} from './buildSteps/dataSetStep';
+import {FilterResultsStep} from './buildSteps/filterResultsStep';
+import {SortByStep} from './buildSteps/sortByStep';
+import {VisualizationStep} from './buildSteps/visualizationStep';
+import {YAxisStep} from './buildSteps/yAxisStep';
 import {Footer} from './footer';
 import {Header} from './header';
-import {SortBySelectors} from './sortBySelectors';
 import {
   DataSet,
   DisplayType,
   getParsedDefaultWidgetQuery,
   mapErrors,
   normalizeQueries,
-  SortDirection,
 } from './utils';
 import {WidgetLibrary} from './widgetLibrary';
-import {YAxisSelector} from './yAxisSelector';
 
 const NEW_DASHBOARD_ID = 'new';
-
-const DATASET_CHOICES: [DataSet, string][] = [
-  [DataSet.EVENTS, t('All Events (Errors and Transactions)')],
-  [DataSet.ISSUES, t('Issues (States, Assignment, Time, etc.)')],
-  // [DataSet.METRICS, t('Metrics (Release Health)')],
-];
 
 function getDataSetQuery(widgetBuilderNewDesign: boolean): Record<DataSet, WidgetQuery> {
   return {
@@ -571,6 +548,15 @@ function WidgetBuilder({
     });
   }
 
+  function handleGetAmendedFieldOptions(measurements: MeasurementCollection) {
+    return generateFieldOptions({
+      organization,
+      tagKeys: Object.values(tags).map(({key}) => key),
+      measurementKeys: Object.values(measurements).map(({key}) => key),
+      spanOperationBreakdownKeys: SPAN_OP_BREAKDOWN_FIELDS,
+    });
+  }
+
   async function dataIsValid(widgetData: Widget): Promise<boolean> {
     if (notDashboardsOrigin) {
       // Validate that a dashboard was selected since api call to /dashboards/widgets/ does not check for dashboard
@@ -683,15 +669,6 @@ function WidgetBuilder({
     });
   }
 
-  function getAmendedFieldOptions(measurements: MeasurementCollection) {
-    return generateFieldOptions({
-      organization,
-      tagKeys: Object.values(tags).map(({key}) => key),
-      measurementKeys: Object.values(measurements).map(({key}) => key),
-      spanOperationBreakdownKeys: SPAN_OP_BREAKDOWN_FIELDS,
-    });
-  }
-
   function isFormInvalid() {
     if (notDashboardsOrigin && !state.selectedDashboard) {
       return true;
@@ -725,7 +702,6 @@ function WidgetBuilder({
   ].includes(state.displayType);
 
   const explodedFields = state.queries[0].fields.map(field => explodeField({field}));
-  const orderBy = state.queries[0].orderby;
 
   return (
     <SentryDocumentTitle title={dashboard.title} orgSlug={orgSlug}>
@@ -749,362 +725,81 @@ function WidgetBuilder({
             <MainWrapper>
               <Main>
                 <BuildSteps symbol="colored-numeric">
-                  <BuildStep
-                    title={t('Choose your visualization')}
-                    description={t(
-                      'This is a preview of how your widget will appear in the dashboard.'
-                    )}
-                  >
-                    <DisplayTypeSelector
-                      displayType={state.displayType}
-                      onChange={(option: {label: string; value: DisplayType}) => {
-                        handleDisplayTypeOrTitleChange('displayType', option.value);
-                      }}
-                      error={state.errors?.displayType}
-                    />
-                    <VisualizationWrapper displayType={state.displayType}>
-                      <WidgetCard
-                        organization={organization}
-                        selection={pageFilters}
-                        widget={currentWidget}
-                        isEditing={false}
-                        widgetLimitReached={false}
-                        renderErrorMessage={errorMessage =>
-                          typeof errorMessage === 'string' && (
-                            <PanelAlert type="error">{errorMessage}</PanelAlert>
-                          )
-                        }
-                        isSorting={false}
-                        currentWidgetDragging={false}
-                        noLazyLoad
-                      />
-                    </VisualizationWrapper>
-                  </BuildStep>
-                  <BuildStep
-                    title={t('Choose your data set')}
-                    description={t(
-                      'This reflects the type of information you want to use. For a full list, read the docs.'
-                    )}
-                  >
-                    <DataSetChoices
-                      label="dataSet"
-                      value={state.dataSet}
-                      choices={DATASET_CHOICES}
-                      disabledChoices={
-                        state.displayType !== DisplayType.TABLE
-                          ? [
-                              [
-                                DataSet.ISSUES,
-                                t(
-                                  'This data set is restricted to the table visualization.'
-                                ),
-                              ],
-                            ]
-                          : undefined
-                      }
-                      onChange={handleDataSetChange}
-                    />
-                  </BuildStep>
+                  <VisualizationStep
+                    widget={currentWidget}
+                    organization={organization}
+                    pageFilters={pageFilters}
+                    displayType={state.displayType}
+                    error={state.errors?.displayType}
+                    onChange={newDisplayType => {
+                      handleDisplayTypeOrTitleChange('displayType', newDisplayType);
+                    }}
+                  />
+                  <DataSetStep
+                    dataSet={state.dataSet}
+                    displayType={state.displayType}
+                    onChange={handleDataSetChange}
+                  />
                   {[DisplayType.TABLE, DisplayType.TOP_N].includes(state.displayType) && (
-                    <BuildStep
-                      title={t('Choose your columns')}
-                      description={
-                        state.dataSet !== DataSet.ISSUES
-                          ? tct(
-                              'To group events, add [functionLink: functions] f(x) that may take in additional parameters. [tagFieldLink: Tag and field] columns will help you view more details about the events (i.e. title).',
-                              {
-                                functionLink: (
-                                  <ExternalLink href="https://docs.sentry.io/product/discover-queries/query-builder/#filter-by-table-columns" />
-                                ),
-                                tagFieldLink: (
-                                  <ExternalLink href="https://docs.sentry.io/product/sentry-basics/search/searchable-properties/#event-properties" />
-                                ),
-                              }
-                            )
-                          : tct(
-                              '[tagFieldLink: Tag and field] columns will help you view more details about the issues (i.e. title).',
-                              {
-                                tagFieldLink: (
-                                  <ExternalLink href="https://docs.sentry.io/product/sentry-basics/search/searchable-properties/#event-properties" />
-                                ),
-                              }
-                            )
-                      }
-                    >
-                      {state.dataSet === DataSet.EVENTS ? (
-                        <Measurements>
-                          {({measurements}) => (
-                            <ColumnFields
-                              displayType={state.displayType}
-                              organization={organization}
-                              widgetType={widgetType}
-                              columns={explodedFields}
-                              errors={state.errors?.queries}
-                              fieldOptions={getAmendedFieldOptions(measurements)}
-                              onChange={handleYAxisOrColumnFieldChange}
-                            />
-                          )}
-                        </Measurements>
-                      ) : (
-                        <ColumnFields
-                          displayType={state.displayType}
-                          organization={organization}
-                          widgetType={widgetType}
-                          columns={state.queries[0].fields.map(field =>
-                            explodeField({field})
-                          )}
-                          errors={
-                            state.errors?.queries?.[0]
-                              ? [state.errors?.queries?.[0]]
-                              : undefined
-                          }
-                          fieldOptions={generateIssueWidgetFieldOptions()}
-                          onChange={newFields => {
-                            const fieldStrings = newFields.map(generateFieldAsString);
-                            const newQuery = cloneDeep(state.queries[0]);
-                            newQuery.fields = fieldStrings;
-                            const {columns, aggregates} =
-                              getColumnsAndAggregates(fieldStrings);
-                            newQuery.aggregates = aggregates;
-                            newQuery.columns = columns;
-                            handleQueryChange(0, newQuery);
-                          }}
-                        />
-                      )}
-                    </BuildStep>
+                    <ColumnsStep
+                      dataSet={state.dataSet}
+                      queries={state.queries}
+                      displayType={state.displayType}
+                      organization={organization}
+                      widgetType={widgetType}
+                      queryErrors={state.errors?.queries}
+                      onQueryChange={handleQueryChange}
+                      onYAxisOrColumnFieldChange={handleYAxisOrColumnFieldChange}
+                      explodedFields={explodedFields}
+                      onGetAmendedFieldOptions={handleGetAmendedFieldOptions}
+                    />
                   )}
                   {![DisplayType.TABLE].includes(state.displayType) && (
-                    <BuildStep
-                      title={
-                        displayType === DisplayType.BIG_NUMBER
-                          ? t('Choose what to plot')
-                          : t('Choose what to plot in the y-axis')
-                      }
-                      description={
-                        [DisplayType.AREA, DisplayType.BAR, DisplayType.LINE].includes(
-                          displayType
-                        )
-                          ? t(
-                              "This is the data you'd be visualizing in the display. You can chart multiple overlays if they share a similar unit."
-                            )
-                          : t("This is the data you'd be visualizing in the display.")
-                      }
-                    >
-                      <Measurements>
-                        {({measurements}) => (
-                          <YAxisSelector
-                            widgetType={widgetType}
-                            displayType={state.displayType}
-                            fields={explodedFields}
-                            fieldOptions={getAmendedFieldOptions(measurements)}
-                            onChange={handleYAxisOrColumnFieldChange}
-                            errors={state.errors?.queries}
-                          />
-                        )}
-                      </Measurements>
-                    </BuildStep>
+                    <YAxisStep
+                      displayType={state.displayType}
+                      widgetType={widgetType}
+                      queryErrors={state.errors?.queries}
+                      onYAxisOrColumnFieldChange={handleYAxisOrColumnFieldChange}
+                      explodedFields={explodedFields}
+                      onGetAmendedFieldOptions={handleGetAmendedFieldOptions}
+                    />
                   )}
-                  <BuildStep
-                    title={t('Filter your results')}
-                    description={
-                      canAddSearchConditions
-                        ? t(
-                            'This is how you filter down your search. You can add multiple queries to compare data.'
-                          )
-                        : t('This is how you filter down your search.')
-                    }
-                  >
-                    <div>
-                      {state.queries.map((query, queryIndex) => {
-                        return (
-                          <QueryField
-                            key={queryIndex}
-                            inline={false}
-                            flexibleControlStateSize
-                            stacked
-                            error={state.errors?.queries?.[queryIndex]?.conditions}
-                          >
-                            <SearchConditionsWrapper>
-                              <Search
-                                searchSource="widget_builder"
-                                organization={organization}
-                                projectIds={selection.projects}
-                                query={query.conditions}
-                                fields={[]}
-                                onSearch={field => {
-                                  // SearchBar will call handlers for both onSearch and onBlur
-                                  // when selecting a value from the autocomplete dropdown. This can
-                                  // cause state issues for the search bar in our use case. To prevent
-                                  // this, we set a timer in our onSearch handler to block our onBlur
-                                  // handler from firing if it is within 200ms, ie from clicking an
-                                  // autocomplete value.
-                                  setBlurTimeout(
-                                    window.setTimeout(() => {
-                                      setBlurTimeout(null);
-                                    }, 200)
-                                  );
-
-                                  const newQuery: WidgetQuery = {
-                                    ...state.queries[queryIndex],
-                                    conditions: field,
-                                  };
-                                  handleQueryChange(queryIndex, newQuery);
-                                }}
-                                onBlur={field => {
-                                  if (!blurTimeout) {
-                                    const newQuery: WidgetQuery = {
-                                      ...state.queries[queryIndex],
-                                      conditions: field,
-                                    };
-                                    handleQueryChange(queryIndex, newQuery);
-                                  }
-                                }}
-                                useFormWrapper={false}
-                                maxQueryLength={MAX_QUERY_LENGTH}
-                              />
-                              {!hideLegendAlias && (
-                                <LegendAliasInput
-                                  type="text"
-                                  name="name"
-                                  value={query.name}
-                                  placeholder={t('Legend Alias')}
-                                  onChange={event => {
-                                    const newQuery: WidgetQuery = {
-                                      ...state.queries[queryIndex],
-                                      name: event.target.value,
-                                    };
-                                    handleQueryChange(queryIndex, newQuery);
-                                  }}
-                                />
-                              )}
-                              {state.queries.length > 1 && (
-                                <Button
-                                  size="zero"
-                                  borderless
-                                  onClick={() => handleQueryRemove(queryIndex)}
-                                  icon={<IconDelete />}
-                                  title={t('Remove query')}
-                                  aria-label={t('Remove query')}
-                                />
-                              )}
-                            </SearchConditionsWrapper>
-                          </QueryField>
-                        );
-                      })}
-                      {canAddSearchConditions && (
-                        <Button
-                          size="small"
-                          icon={<IconAdd isCircled />}
-                          onClick={handleAddSearchConditions}
-                        >
-                          {t('Add query')}
-                        </Button>
-                      )}
-                    </div>
-                  </BuildStep>
+                  <FilterResultsStep
+                    queries={state.queries}
+                    hideLegendAlias={hideLegendAlias}
+                    canAddSearchConditions={canAddSearchConditions}
+                    organization={organization}
+                    onSetBlurTimeout={setBlurTimeout}
+                    blurTimeout={blurTimeout}
+                    queryErrors={state.errors?.queries}
+                    onAddSearchConditions={handleAddSearchConditions}
+                    onQueryChange={handleQueryChange}
+                    onQueryRemove={handleQueryRemove}
+                  />
                   {[DisplayType.TABLE, DisplayType.TOP_N].includes(state.displayType) && (
-                    <BuildStep
-                      title={t('Sort by a column')}
-                      description={t(
-                        "Choose one of the columns you've created to sort by."
-                      )}
-                    >
-                      <Field
-                        inline={false}
-                        error={state.errors?.orderby}
-                        flexibleControlStateSize
-                        stacked
-                      >
-                        {widgetBuilderNewDesign ? (
-                          <SortBySelectors
-                            sortByOptions={
-                              state.dataSet === DataSet.EVENTS
-                                ? generateOrderOptions({
-                                    widgetType,
-                                    widgetBuilderNewDesign: true,
-                                    ...getColumnsAndAggregates(state.queries[0].fields),
-                                  })
-                                : generateIssueWidgetOrderOptions(
-                                    organization.features.includes(
-                                      'issue-list-trend-sort'
-                                    )
-                                  )
-                            }
-                            values={{
-                              sortDirection:
-                                orderBy[0] === '-'
-                                  ? SortDirection.HIGH_TO_LOW
-                                  : SortDirection.LOW_TO_HIGH,
-                              sortBy:
-                                orderBy[0] === '-'
-                                  ? orderBy.substring(1, orderBy.length)
-                                  : orderBy,
-                            }}
-                            onChange={({sortDirection, sortBy}) => {
-                              const newQuery: WidgetQuery = {
-                                ...state.queries[0],
-                                orderby:
-                                  sortDirection === SortDirection.HIGH_TO_LOW
-                                    ? `-${sortBy}`
-                                    : sortBy,
-                              };
-                              handleQueryChange(0, newQuery);
-                            }}
-                          />
-                        ) : (
-                          <SelectControl
-                            menuPlacement="auto"
-                            value={
-                              state.dataSet === DataSet.EVENTS
-                                ? state.queries[0].orderby
-                                : state.queries[0].orderby || IssueSortOptions.DATE
-                            }
-                            name="orderby"
-                            options={
-                              state.dataSet === DataSet.EVENTS
-                                ? generateOrderOptions({
-                                    widgetType,
-                                    ...getColumnsAndAggregates(state.queries[0].fields),
-                                  })
-                                : generateIssueWidgetOrderOptions(
-                                    organization.features.includes(
-                                      'issue-list-trend-sort'
-                                    )
-                                  )
-                            }
-                            onChange={(option: SelectValue<string>) => {
-                              const newQuery: WidgetQuery = {
-                                ...state.queries[0],
-                                orderby: option.value,
-                              };
-                              handleQueryChange(0, newQuery);
-                            }}
-                          />
-                        )}
-                      </Field>
-                    </BuildStep>
+                    <SortByStep
+                      queries={state.queries}
+                      dataSet={state.dataSet}
+                      widgetBuilderNewDesign={widgetBuilderNewDesign}
+                      error={state.errors?.orderby}
+                      onQueryChange={handleQueryChange}
+                      organization={organization}
+                      widgetType={widgetType}
+                    />
                   )}
                   {notDashboardsOrigin && (
-                    <BuildStep
-                      title={t('Choose your dashboard')}
-                      description={t(
-                        "Choose which dashboard you'd like to add this query to. It will appear as a widget."
-                      )}
-                      required
-                    >
-                      <DashboardSelector
-                        error={state.errors?.dashboard}
-                        dashboards={state.dashboards}
-                        onChange={selectedDashboard =>
-                          setState({
-                            ...state,
-                            selectedDashboard,
-                            errors: {...state.errors, dashboard: undefined},
-                          })
-                        }
-                        disabled={state.loading}
-                      />
-                    </BuildStep>
+                    <DashboardStep
+                      error={state.errors?.dashboard}
+                      dashboards={state.dashboards}
+                      onChange={selectedDashboard =>
+                        setState({
+                          ...state,
+                          selectedDashboard,
+                          errors: {...state.errors, dashboard: undefined},
+                        })
+                      }
+                      disabled={state.loading}
+                    />
                   )}
                 </BuildSteps>
               </Main>
@@ -1142,38 +837,6 @@ export default withPageFilters(withTags(WidgetBuilder));
 
 const PageContentWithoutPadding = styled(PageContent)`
   padding: 0;
-`;
-
-const VisualizationWrapper = styled('div')<{displayType: DisplayType}>`
-  overflow: ${p => (p.displayType === DisplayType.TABLE ? 'hidden' : 'visible')};
-  padding-right: ${space(2)};
-`;
-
-const DataSetChoices = styled(RadioGroup)`
-  @media (min-width: ${p => p.theme.breakpoints[2]}) {
-    grid-auto-flow: column;
-  }
-`;
-
-const SearchConditionsWrapper = styled('div')`
-  display: flex;
-  align-items: center;
-
-  > * + * {
-    margin-left: ${space(1)};
-  }
-`;
-
-const Search = styled(SearchBar)`
-  flex-grow: 1;
-`;
-
-const LegendAliasInput = styled(Input)`
-  width: 33%;
-`;
-
-const QueryField = styled(Field)`
-  padding-bottom: ${space(1)};
 `;
 
 const BuildSteps = styled(List)`
